+++ src/ut/ut_rvaluereference/usage/ncOldCustomString.h
@@ -34,7 +34,7 @@ public:
         
         _str = new char[strlen(source)+1];
         strcpy(_str, source);
-        usleep (LOSS_TIME * strlen(source) );//模拟其他数据成员的memcopy
+        usleep (LOSS_TIME * (int)strlen(source) );//模拟其他数据成员的memcopy
     }
     
     ncOldCustomString(const ncOldCustomString& source)
@@ -43,7 +43,7 @@ public:
         
         _str = new char[strlen(source._str) + 1];
         strcpy(_str, source._str);
-        usleep (LOSS_TIME * strlen(source._str) );//模拟其他数据成员的memcopy
+        usleep (LOSS_TIME * (int)strlen(source._str) );//模拟其他数据成员的memcopy
     }
     
     ncOldCustomString& operator = (const ncOldCustomString& source)
@@ -53,7 +53,7 @@ public:
         if (this != &source){
             char *tmp = new char[strlen(source._str) + 1];
             strcpy(tmp, source._str);
-            usleep (LOSS_TIME * strlen(source._str) );//模拟其他数据成员的memcopy
+            usleep (LOSS_TIME * (int)strlen(source._str) );//模拟其他数据成员的memcopy
             delete []_str;
             _str = tmp;
         }
